requirements.txt
rich
tefas-crawler
